FE: move method into separate file
